(refs #1101)Fix Batch API url mapping
